make lyricspy.aio the default backend
doing the same thing twice is boring
