+++ lyricspy/__init__.py
@@ -5,16 +5,18 @@ from typing import Union
 import httpx
 from bs4 import BeautifulSoup
 
+from asyncio import get_event_loop
+
 headers = {
     'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.163 Safari/537.37'}
 class Musixmatch:
     def __init__(self, usertoken:Union[str, list]):
         self.token = usertoken
-        self.http = httpx.Client(http2=True)
+        self.http = httpx.AsyncClient(http2=True)
 
-    def search(self, q, limit):
+    async def search(self, q, limit):
         utoken = random.choice(self.token) if type(self.token) is list else self.token
-        a = self.http.get('https://apic.musixmatch.com/ws/1.1/macro.search', params=dict(
+        a = await self.http.get('https://apic.musixmatch.com/ws/1.1/macro.search', params=dict(
             app_id='android-player-v1.0',
             usertoken=utoken,
             q=q,
@@ -24,22 +26,32 @@ class Musixmatch:
         ), headers=headers)
         return a.json()
 
-    def lyrics(self, id=None, artist=None, track=None):
+    async def lyrics(self, id):
         utoken = random.choice(self.token) if type(self.token) is list else self.token
-        a = self.http.get('https://apic.musixmatch.com/ws/1.1/macro.subtitles.get', params=dict(
+        a = await self.http.get('https://apic.musixmatch.com/ws/1.1/macro.subtitles.get', params=dict(
             app_id='android-player-v1.0',
             usertoken=utoken,
-            q_artist=artist,
-            q_track=track,
             track_id=id,
             format='json'
         ), headers=headers, follow_redirects=True)
 
         return a.json()
 
-    def translation(self, id, lang):
+    async def spotify_lyrics(self, artist, track):
         utoken = random.choice(self.token) if type(self.token) is list else self.token
-        a = self.http.get('https://apic.musixmatch.com/ws/1.1/crowd.track.translations.get', params=dict(
+        a = await self.http.get('https://apic.musixmatch.com/ws/1.1/macro.subtitles.get', params=dict(
+            app_id='android-player-v1.0',
+            usertoken=utoken,
+            q_artist=artist,
+            q_track=track,
+            format='json'
+        ), headers=headers, follow_redirects=True)
+
+        return a.json()
+
+    async def translation(self, id, lang, letra=None):
+        utoken = random.choice(self.token) if type(self.token) is list else self.token
+        a = await self.http.get('https://apic.musixmatch.com/ws/1.1/crowd.track.translations.get', params=dict(
             app_id='android-player-v1.0',
             usertoken=utoken,
             selected_language=lang,
@@ -47,23 +59,35 @@ class Musixmatch:
             part='user',
             format='json'
         ), headers=headers)
-        return a.json()
-
-    def auto(self, q, lang, limit=5):
-        print(limit)
-        a = self.search(q, limit)
-        res = a['message']['body']['macro_result_list']['track_list'] if limit != 1 else [a['message']['body']['macro_result_list']['best_match']]
-        ret = []
-        for i in res:
-            id = i['track']['track_id'] if 'track' in i else i['id']
-            b = self.lyrics(id)
-            letra = b['message']['body']['macro_calls']['track.lyrics.get']['message']['body']['lyrics']['lyrics_body']
-            c = self.translation(id, lang)
+        print(a.url)
+        c = a.json()
+        if c['message']['body']['translations_list'] and letra:
             tr = letra
             for i in c['message']['body']['translations_list']:
                 escape = re.escape(i['translation']['snippet'])
                 tr = re.sub(f'^{escape}$', i['translation']['description'], tr, flags=re.M)
-            b['translate'] = tr
+        elif c['message']['body']['translations_list']:
+            tr = True
+        else:
+            tr = None
+
+        return tr
+
+    async def auto(self, q=None, lang="pt", limit=5, id=None):
+        print("auto", q, lang, limit, id)
+        if q:
+            a = await self.search(q, limit)
+            res = a['message']['body']['macro_result_list']['track_list'] if limit != 1 else [a['message']['body']['macro_result_list']['best_match']]
+        else:
+            res = [id]
+        ret = []
+        for i in res:
+            if not id:
+                id = i['track']['track_id'] if 'track' in i else i['id']
+            b = await self.lyrics(id)
+            letra = b['message']['body']['macro_calls']['track.lyrics.get']['message']['body']['lyrics']['lyrics_body']
+            c = await self.translation(id, lang, letra)
+            b['translate'] = c
             ret.append(b)
         return ret
 
@@ -79,11 +103,11 @@ class Musixmatch:
 
 class Letras:
     def __init__(self):
-        self.http = httpx.Client(http2=True)
+        self.http = httpx.AsyncClient(http2=True)
 
-    def letra(self, query, **kwargs):
+    async def letra(self, query, **kwargs):
         link = query['link'].replace('www.letras', 'm.letras')
-        r = self.http.get(link, params=dict(**kwargs))
+        r = await self.http.get(link, params=dict(**kwargs))
         tr = None
         soup = BeautifulSoup(r.text, "html.parser")
         for br in soup.find_all("br"):
@@ -102,8 +126,8 @@ class Letras:
             query.update({'traducao':None})
         return query
 
-    def search(self, query):
-        r = self.http.get('https://studiosolsolr-a.akamaihd.net/letras/app2/', params=dict(q=query))
+    async def search(self, query):
+        r = await self.http.get('https://studiosolsolr-a.akamaihd.net/letras/app2/', params=dict(q=query))
         a = r.json()
         x = []
         n = 0
@@ -115,11 +139,11 @@ class Letras:
             n += 1
         return x
 
-    def auto(self, query, limit=4):
+    async def auto(self, query, limit=4):
         result = []
         n = 0
-        for i in self.search(query):
-            a = self.letra(i)
+        for i in await self.search(query):
+            a = await self.letra(i)
             result.append(a)
             n += 1
             if n == limit:
